requirements.txt
pandas==2.1.0
snowflake-connector-python==3.5.0
snowflake-sqlalchemy==1.5.0
python-dotenv==1.0.0
tqdm==4.66.0